backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.9.0
anthropic==0.89.0
requests==2.33.1
beautifulsoup4==4.14.3
python-dotenv==1.2.2
openpyxl==3.1.5
